+++ js/utils/uiUtils/common/buttonUtils.js
@@ -1,0 +1,11 @@
+export const initialButtons = (buttonMessages, buttonActionMapper) => {
+  const buttonContainer = document.getElementById("buttonContainer");
+  Object.entries(buttonMessages).forEach(([key, name]) => {
+    const button = document.createElement("div");
+    button.className = "button-item";
+    button.textContent = name; // 버튼에 표시할 텍스트 설정
+
+    button.addEventListener("click", buttonActionMapper[key]);
+    buttonContainer.appendChild(button); // 버튼을 buttonContainer에 추가
+  });
+}
